refactor(react): wip, intermediate step renaming index.d.ts to index-types.ts, minor fixes
