+++ packages/botonic-react/src/components/index.ts
@@ -7,6 +7,7 @@ export { Document } from './document'
 export { Element } from './element'
 export { Image } from './image'
 export { Location } from './location'
+export { Message } from './message'
 export { MessageTemplate } from './message-template'
 export * from './multichannel'
 export { Pic } from './pic'
