+++ packages/botonic-react/src/webchat/index.ts
@@ -2,9 +2,6 @@
 export { Webchat } from './webchat'
 export { WebchatDev } from './webchat-dev'
 
-/**
- * @returns {WebChatApp}
- */
 export function getBotonicApp() {
   // Botonic is exported from your bot
   // eslint-disable-next-line no-undef
